Added a simiple rule
This shows the issue at hand the the entire package is too big to
execute on. There are rules that run on just the vehicle with collection
of coverages and some that run on the entire policy. Therefore there
must be a difference. and they should be executed as such.  So I can see
the entire container executing the same rule set for each vehicle.